requirements.txt
# requirements.txt
pandas>=1.3.0
yfinance>=0.2.0
requests>=2.25.0
matplotlib>=3.10.3
python-dotenv>=1.1.1
boto3>=1.39.13
jinja2>=3.1.6
PyYAML>=6.0.3
Flask>=3.0.0
flask-cors>=4.0.0
fastmcp>=2.0.0
transformers>=4.30.0
torch>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
google-cloud-secret-manager>=2.20.0
cloud-sql-python-connector[psycopg2]>=1.10.0
google-auth>=2.29.0
PyJWT>=2.8.0
google-cloud-pubsub>=2.21.0
gunicorn>=21.2.0
